Core: Order update hooks to avoid conflicts
During rollout of release 7.x-6.6.0-beta5 we are seeing warnings
due to missing columns in the user table when accessed by the
entity cache. This is happening when running update hook 
ding2_update_7102 where importing translations also clears the cache.

To avoid this we mark ding2_update_7102 as being dependent on
user_update_7020 - the update hook responsible for adding the 
additional columns.