Push removeIdentifierOptOut when the user does not accept cookies
This has been requested by Mapp support.

We also remove pushing setIdentifierOptOut when the script loads if
consent is not given. This has not explicitly been requested but it does
not seem to make sense to do this as it means that setIdentifierOptOut
will be pushed on the next pageload after consent has been withdrawn.